requirements.txt
pydub~=0.24.1  # pydub requires ffmpeg
speechrecognition~=3.8.1
pyaudio~=0.2.11
sounddevice~=0.4.1
soundfile~=0.10.3.post1
numpy~=1.19.3
pysimplegui~=4.32.1